backend/audio-processing/requirements.txt
fastapi
uvicorn[standard]
pydantic
grpcio
protobuf
openai-whisper
redis
kafka-python
python-multipart
